+++ cmd/kcm/main.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"errors"
 	"fmt"
 	"os"
 
@@ -11,7 +12,13 @@ var version = "0.1.0-dev"
 
 func main() {
 	if err := kcm.NewCommand(version).Execute(); err != nil {
-		fmt.Fprintln(os.Stderr, "kcm:", err)
+		if err.Error() != "" {
+			fmt.Fprintln(os.Stderr, "kcm:", err)
+		}
+		var status *kcm.ExitError
+		if errors.As(err, &status) {
+			os.Exit(status.Code)
+		}
 		os.Exit(1)
 	}
 }